requirements.txt
fastapi
uvicorn
pydantic
langchain
langgraph
openai
yt-dlp
python-dotenv
playwright
edge-tts
